backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
asttokens==3.0.0
bcrypt==5.0.0
beautifulsoup4==4.14.2
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
decorator==5.2.1
Deprecated==1.3.1
docopt==0.6.2
executing==2.2.1
fastapi==0.128.7
h11==0.16.0
idna==3.11
ipython==9.7.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
jsonpickle==4.1.1
limits==5.8.0
MarkupSafe==3.0.3
matplotlib-inline==0.2.1
networkx==3.5
numpy==2.3.4
packaging==26.0
pandas==2.3.3
parso==0.8.5
pipreqs==0.4.13
prompt_toolkit==3.0.52
psycopg2-binary==2.9.11
pure_eval==0.2.3
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.22
pytz==2025.2
pyvis==0.3.2
requests==2.32.5
six==1.17.0
slowapi==0.1.9
soupsieve==2.8
stack-data==0.6.3
starlette==0.52.1
traitlets==5.14.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.40.0
wcwidth==0.2.14
wrapt==2.1.1
yarg==0.1.10
slowapi
